+++ rust/src/lib.rs
@@ -225,7 +225,7 @@ pub async fn proto_find_closest_ticker_ids(ticker_id: TickerId) -> Result<JsValu
         // Convert original PCA coordinates to JS array
         let original_pca_array = js_sys::Array::new();
         for coord in original_pca_coords {
-            original_pca_array.push(&JsValue::from_f64(coord));
+            original_pca_array.push(&JsValue::from_f64(coord as f64));
         }
         js_sys::Reflect::set(
             &obj,
@@ -237,7 +237,7 @@ pub async fn proto_find_closest_ticker_ids(ticker_id: TickerId) -> Result<JsValu
         // Convert translated PCA coordinates to JS array
         let translated_pca_array = js_sys::Array::new();
         for coord in translated_pca_coords {
-            translated_pca_array.push(&JsValue::from_f64(coord));
+            translated_pca_array.push(&JsValue::from_f64(coord as f64));
         }
         js_sys::Reflect::set(
             &obj,
